feat: extract LocalRunnersView from SettingsView (#1202)
Moves all local runner management UI and logic out of the monolithic
SettingsView into a new dedicated LocalRunnersView, accessible via a
"Manage local runners" navigation row in Settings.

Changes:
- Add Sources/RunnerBar/Views/Settings/LocalRunnersView.swift
  Owns all runner-specific @State, @StateObject, lifecycle actions
  (performResume, performStop, performRemoval), row builders, popover
  wiring, and AddRunnerSheet/RemovalAlertModifier presentation.

- Update Sources/RunnerBar/Views/Settings/SettingsView.swift
  Replace the ~250-line localRunnersSection + all runner-specific state
  with a showLocalRunners @State flag and a manageLocalRunnersRow nav
  entry. Body conditionally swaps to LocalRunnersView via flag-swap,
  consistent with the existing panel nav model (no NavigationStack).
  HEIGHT/WIDTH CONTRACT comments preserved verbatim.

SwiftLint: file_header, missing_docs, sorted_imports, closure_spacing all
satisfied. Periphery: LocalRunnersView is referenced from SettingsView so
no orphaned symbols. Tests: no model/store changes; RunnerBarCoreTests and
RunnerBarUITests are unaffected.

Closes #1202
Ref #1206
